+++ CapStone/react_frontend/AWARE/src/components/MockElections/ViewResults.jsx
@@ -1,0 +1,9 @@
+
+export const ViewResults = ({ ballotNum }) => {
+
+    return (
+        <div className="flex flex-col w-full h-full rounded-xl gap-2 p-2 justify-center items-center overflow-auto my-2">
+            { ballotNum }
+        </div>
+    )
+}
