+++ CapStone/react_frontend/AWARE/src/pages/Student/Elections.jsx
@@ -1,13 +1,16 @@
 import { useState, useEffect } from "react";
 import { ToastContainer, toast } from 'react-toastify';
-import { StudentElectionCard, VoteOnBallot } from "../../components/MockElections/StudentElectionCards";
+import { StudentElectionCard} from "../../components/MockElections/StudentElectionCards";
+import { ViewResults} from "../../components/MockElections/ViewResults";
+import { VoteOnBallot } from "../../components/MockElections/Vote";
 
 
 function Elections() {
 
     const [elections, setElections] = useState([]);
     const [voting, setVoting] = useState(false);
-    const [votingOn, setVotingOn] = useState();
+    const [currentBallot, setCurrentBallot] = useState();
+    const [viewingResults, setViewingResults] = useState(false);
 
     useEffect(() => {
         const user = JSON.parse(localStorage.getItem('user'));
@@ -26,22 +29,29 @@ function Elections() {
     }, []);
 
     function onVote(key) {
-        setVotingOn(key-1);
+        setCurrentBallot(key-1);
         setVoting(true);
     }
 
-    function onViewResults() {
-
+    function onViewResults(key) {
+        setCurrentBallot(key-1)
+        setViewingResults(!viewingResults);
     }
 
-    const availableQuizzesView = `${voting ? 'hidden' : 'flex flex-col relative h-[85%] w-[95%] rounded-xl items-center justify-center bg-white mb-4 shadow-lg'}`
+    const availableQuizzesView = `${voting || viewingResults ? 'hidden' : 'flex flex-col relative h-[85%] w-[95%] rounded-xl items-center justify-center bg-white mb-4 shadow-lg'}`
 
     return (
         <div className='flex flex-col h-[91vh] w-[100%] bg-slate-400 justify-center items-center'>
             <div className='flex flex-col h-[100%] w-[90%] rounded-xl items-center'>
                 <div className='flex flex-row h-[15%] w-[95%] items-center justify-center'>
                     <div className='text-3xl md:text-4xl lg:text-5xl text-white drop-shadow-lg'>
-                        {voting ? <h1>{elections[votingOn]['electionTitle']}</h1> : 'Ballots'}
+                        {voting || viewingResults? 
+                            <h1>
+                                {elections[currentBallot]['electionTitle']}
+                            </h1> 
+                        : 
+                            'Ballots'
+                        }
                     </div>
                 </div>
                 <div className={availableQuizzesView}>
@@ -83,11 +93,16 @@ function Elections() {
                 {voting ? 
                     <div className="flex flex-col relative h-[85%] w-[95%] rounded-xl items-center justify-center bg-white mb-4 shadow-lg">
                         <VoteOnBallot 
-                            ballotNum = {elections[votingOn]['ballotNum']}
+                            ballotNum = {elections[currentBallot]['ballotNum']}
                         /> 
                     </div>
                 : 
-                    ''
+                    viewingResults ? 
+                        <div className="flex flex-col relative h-[85%] w-[95%] rounded-xl items-center justify-center bg-white mb-4 shadow-lg">
+                            <ViewResults ballotNum={elections[currentBallot]['ballotNum']}/>
+                        </div> 
+                    :
+                        ''
                 }
             </div>
             <ToastContainer
